Se cambio la posicion de los tests relacionados con el estado de un evento
